+++ src/fsm/base_explore_fsm.cpp
@@ -39,7 +39,7 @@ NS_START(fordyca, fsm);
  * Constructors/Destructors
  ******************************************************************************/
 base_explore_fsm::base_explore_fsm(
-    double unsuccessful_dir_change_thresh,
+    uint unsuccessful_dir_change_thresh,
     const std::shared_ptr<rcppsw::er::server> &server,
     const std::shared_ptr<controller::base_foraging_sensors> &sensors,
     const std::shared_ptr<controller::actuator_manager> &actuators,
